fix exponent

+++ proto/cabi.thrift
@@ -19,7 +19,7 @@ enum ExchangeAction {
 
 struct Currency {
     1: required CurrencySymbolicCode symbolic_code
-    2: required i16 exponent
+    2: required i32 exponent
 }
 
 struct CheckCurrencyExchangeParams {
